Some Minor Output and Error Clean up for netperf (#4072)

+++ scripts/secnetperf.ps1
@@ -108,7 +108,9 @@ if ($io -eq "wsk") {
 # Copy the artifacts to the peer.
 Write-Host "Copying files to peer"
 Invoke-Command -Session $Session -ScriptBlock {
-    Remove-Item -Force -Recurse $Using:RemoteDir | Out-Null
+    if (Test-Path $Using:RemoteDir) {
+        Remove-Item -Force -Recurse $Using:RemoteDir | Out-Null
+    }
     mkdir $Using:RemoteDir | Out-Null
 }
 Copy-Item -ToSession $Session ./artifacts -Destination "$RemoteDir/artifacts" -Recurse
@@ -243,6 +245,9 @@ Write-Host "Tests complete!"
     Write-Host "`Writing json-test-results-$plat-$os-$arch-$tls-$io.json"
     $json | ConvertTo-Json | Set-Content -Path "json-test-results-$plat-$os-$arch-$tls-$io.json"
 }
+
+# Clear out any exit codes from previous commands.
+$global:LastExitCode = 0
 
 if ($hasFailures) {
     exit 1
